Update to v1.6.0 of the TF provider
This bumps our version to v1.6.0 of the TF provider.  This picks up some
bug fixes that are blocking scenarios I'm looking at porting.  In
general, we need to get better at updating to the latest more often.

Note that we're building up a backlog of unmapped resources also.  I've
not yet addressed those, but we should look as part of pulumi/home#265.