Depend on @pulumi/pulumi directly
Instead of generating a package.json that has `@pulumi/pulumi` as a
peerDependency and then managing the promotion of it into an actual
dependency. Just express it as an actual dependency in the geneated
package.json file.

In addition, we stop linking `@pulumi/pulumi` in our tests, instead
having an actual dependency in their package.json's that we lock to.

This means we no longer require the `install-locked.sh` dance in CI to
download a prebuilt copy of @pulumi/pulumi and register it with yarn
link.