Update Gopkg.toml using govendor-override.
These changes add explicit overrides for the transitive dependencies of
`terraform-provider-azurerm` at the versions specified in that project's
`vendor.json` file. This ensures that we will build using package
versions with which the TF provider is known to work.

These changes also update the referenced version of the TF provider.