Don't auto-name when 'name' is optional
We have been applying auto-naming to all input properties, even if they were already Optional.  In that case, the provider already understands how to provide a default name - or one is truly not required at all.  So we should avoid overwriting this with a Pulumi autoname.

Also fixes the casing of `azure.role.assignment` - this is a breaking change.