Fix package install on Windows
Our README.rst (which is generated from README.md) has smart quotes in
it, which can break install of the package on Windows (since reading
the README.rst fails because it is not valid in the current windows
code-page).

This change ensures we don't generate such `.rst` files. This is the
quickest thing to to address the issue - a longer term fix I'll land
everywhere soon is to just always open these files as UTF-8, and
hopefully remove the need to use `.rst` files at all.

Fixes: #356