Avoid storing config from the environment into the state.
Like in https://github.com/pulumi/pulumi-aws/pull/894, I am not sure this is the right way to fix this long term, but it is a necessary fix to ensure secrets are not captured from the environment and stored in the state file.

This is a breaking change for use of `azure.config.*` which will no longer get populated from ambient environment variables.

I have left `subscription_id`, `environment` and `location` as-is as those are all immutable aspects of the provider and contain non-sensitive information.  I'm not yet sure the precise principled answer to what should be included vs. not here.

Fixes #576.