Fix creation of azure.compute.*VirtualMachine (#881)
Fixes: #860

A change was merged to the upstream provider that introduced a
RequiredWith check. At the same time, a default value of `-1` was
added to the `platformFaultDomain` value. This means that when a
Linux / Windows Virtual Machine was tried to be provisioned, it was
failing the `RequiredWith` check as `-1` was being passed as a value

The bug fix here was to relax the `RequiredWith` check. Also included
a reproduction that will now succeed that failed previously as follows:

```
  azure:compute:LinuxVirtualMachine (exampleLinuxVirtualMachine):
    error: azure:compute/linuxVirtualMachine:LinuxVirtualMachine resource 'exampleLinuxVirtualMachine' has a problem: "platform_fault_domain": all of `platform_fault_domain,virtual_machine_scale_set_id` must be specified
```