Reinstate plugin version bump in Makefile
The Makefile here is generated, and pushed, by
https://github.com/pulumi/ci-mgmt. However, a change was made locally in

    https://github.com/pulumi/pulumi-azure/pull/1197

The correct but circuitous way to fix this, while these files are under
control of ci-mgmt, is to change it there and regenerate the files, then
push the generated files downstream. As a mitigation to get the build
working, this commit reinstates the local change so that merging this
automated PR doesn't revert it.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>