Upgrade test validate no replacements (#2132)
The upgrade tests here were using the no-diff validation yielding noise
in the tests as can be seen here:
https://github.com/pulumi/pulumi-azure/actions/runs/9505545675/job/26201251247?pr=2125

The tests there fail because of default properties applied to storage
account and storage bucket. The azure provider is especially prone to
dirty refreshes because of a policy the upstream provider has AGAINST
setting Optional + Computed for properties returned by the API:
https://github.com/hashicorp/terraform-provider-azurerm/blob/main/contributing/topics/best-practices.md#setting-properties-to-optional--computed

No diff is too strict for upgrade tests so this PR replaces it with "no
replacements".