Upgrade terraform-provider-azurerm to v4.15.0 (#2927)
This PR was generated via `$ upgrade-provider pulumi/pulumi-azure
--kind=all --target-bridge-version=latest`.

## Manual edits

1. Resolved a simple conflict in
`upstream/website/docs/r/api_management_api_policy.html.markdown`.
2. Ignored deprecated resource `azurerm_network_packet_capture` by
adding it to `IgnoreMappings` because it failed the build with _"could
not find docs"_.

(2) is a bit odd because the resource isn't in the current version of
the pulumi provider at all. (Otherwise, we couldn't just remove it.) I
assume that due to upstream [azurerm_network_packet_capture - deprecate
resource for
5.0](https://github.com/hashicorp/terraform-provider-azurerm/pull/28324),
which changed the version gates for the resource, it was newly included
but we don't want it.

---

- Upgrading terraform-provider-azurerm from 4.14.0  to 4.15.0.
	Fixes #2925