run ci-mgmt on v5 (#3345)
This pull request is based on the `v5 branch, a branch created off the
last v5 published version (v5.89.0).

It updates the Workflow files to reflect modern ci-mgmt workflows more
closely.

Notable challenges:

- pulumi-java still needs its own pulumi-java-gen binary for this
version of Pulumi so the Makefile Java generation looks different (old
format).
- There was an update to a Node dependency, `@azure/identity`, that
resulted in needing a post-codegen rewrite of a few Node files (see
commit
https://github.com/pulumi/pulumi-azure/pull/3345/commits/021fd8375b4e1a137706dbe16916a063521c8472).
- There have been some slight updates to the secondary plugins,
resulting in a few documentation-only schema diffs, and SDKs had to be
regenerated to avoid a dirty worktree.
- ci-mgmt.yml was updated to use ESC and to no longer run separate tests
with `gotestfmt`
- I increased the runner size as well.
- Frontdoor test was removed as new creations are no longer allowed due
to deprecation of the resource
- Default VM was updated in Webserver tests


Fast-forward ci-mgmt update for v5 branch.